Finish pytest fixture + tox axis setup
