+++ tox.ini
@@ -1,9 +1,13 @@
 [tox]
-envlist = py27,py35
+envlist = {py27,py35}-unit,{py27,py35}-{functional,integration}-{mockthink,rethink}
 
 [testenv]
-whitelist_externals = make
 deps =
     pytest
     py27: mock
-commands = make test
+commands =
+    unit: py.test mockthink/test/unit {posargs}
+    functional-mockthink: py.test mockthink/test/functional --run mockthink {posargs}
+    functional-rethink: py.test mockthink/test/functional --run rethink {posargs}
+    integration-mockthink: py.test mockthink/test/integration --run mockthink {posargs}
+    integration-rethink: py.test mockthink/test/integration --run rethink {posargs}
